# specs/004-identity/contracts/api.yaml
openapi: 3.0.3
info:
  title: RAGGit Workstation API (Single-Tenant)
  version: 1.3.0
  description: >
    Single-tenant on-prem RAG on AI Workstation (LanceDB local path +
    Ollama/LLamaSharp/ONNX). Desktop is thin client over LAN (no WAN at query
    time per FR-004/Constitution IV). 1.3.0 (004-identity, MINOR/ADDITIVE per
    FR-013): local per-person accounts — POST /api/auth/login and
    POST /api/auth/refresh issue workstation-signed JWT bearer tokens (8h);
    /api/users CRUD (Admin-only) manages the people directory; GET /api/auth/me
    gains ADDITIVE optional fields (displayName, username, sub) — 1.2.0
    required fields unchanged, clients MUST ignore unknown fields; Bearer
    security scheme added alongside retained ApiKey (machine/bootstrap per
    FR-012); credentials and session tokens traverse the LAN over HTTPS only
    (FR-008); 429 added for sign-in lockout (FR-009, 5 attempts / 15 min);
    Documents.createdBy and Queries attribution carry the stable person id for
    Local sessions while legacy "admin"/"employee" values remain valid
    (FR-007). No breaking changes: every 1.2.0 path, schema, and required
    field is preserved. 1.2.0 added xlsx MIME + rejection ordering; 1.1.0
    added GET /api/auth/me (extensible envelope) and pdf/docx 400 hardening.
servers:
  - url: https://ai-workstation.local:5001
    description: AI Workstation LAN — HTTPS required for credential and token traffic (FR-008)
  - url: http://localhost:5001
...
security:
  - ApiKey: []
  - Bearer: []
paths:
...
  /api/users:
    get:
      summary: List person accounts (Admin only)
      description: >
        Returns the people directory. PasswordHash is NEVER serialized.
        Requires Admin — via Local JWT role claim or Admin API key (bootstrap).
        Employee callers receive 403 (US3.3).
      operationId: listUsers
      security:
        - ApiKey: []
        - Bearer: []
      responses:
        '200':
          description: People directory
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      summary: Create a person account (Admin only)
      description: >
        Creates an account with role Admin|Employee and a PBKDF2-hashed
        password (FR-001, FR-005). Takes effect immediately — no restart
        (SC-006). Duplicate username (case-insensitive) returns 409.
      operationId: createUser
      security:
        - ApiKey: []
        - Bearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: Created — account can sign in immediately
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: Username already exists (case-insensitive uniqueness)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/users/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
        description: Stable person id (User.Id)
    get:
      summary: Read one person account (Admin only)
      operationId: getUser
      security:
        - ApiKey: []
        - Bearer: []
      responses:
        '200':
          description: Account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      summary: Change role / deactivate / reactivate / rename (Admin only)
      description: >
        Partial update — only role, isActive, displayName. Deactivation
        (isActive=false) refuses the account's VERY NEXT request on any
        existing session — live tokens are checked against the Users row per
        request (FR-010, SC-003); takes effect without restarting the
        workstation (SC-006). Accounts are never hard-deleted (attribution
        references persist).
      operationId: patchUser
      security:
        - ApiKey: []
        - Bearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPatch'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/users/{id}/reset-password:
    post:
      summary: Reset a person's password (Admin only)
      description: >
        Sets a new password (PBKDF2-hashed) and optionally forces a change at
        next sign-in (mustChangePassword — spec edge case). Clears lockout and
        failure counters. No restart required (SC-006).
      operationId: resetUserPassword
      security:
        - ApiKey: []
        - Bearer: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                password: { type: string, format: password, minLength: 10 }
                mustChangePassword: { type: boolean, default: false }
              required: [password]
      responses:
        '204': { description: Password reset, content: {} }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    ApiKey:
      type: apiKey
      in: header
      name: X-Api-Key
      description: >
        Single-tenant machine/bootstrap key (FR-012). Retained for machine
        integration and initial bring-up; NOT used for per-person attribution
        when a Local session exists (FR-007). Windows AD remains a documented
        future provider.
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Workstation-signed HS256 JWT issued by POST /api/auth/login (004).
        8-hour lifetime; identifies a person (sub = stable person id), role,
        and display name. HTTPS-only transport (FR-008). Deactivated or
        locked-out accounts are refused per request (FR-010).
  schemas:
...
    User:
      type: object
      description: Person account (004). PasswordHash is never serialized. No tenant/company field (Constitution I).
      properties:
        id: { type: string, format: uuid, description: Stable person id — attribution key (FR-007) }
        username: { type: string, example: ada }
        displayName: { type: string, example: Ada Lovelace }
        role: { type: string, enum: [Admin, Employee] }
        isActive: { type: boolean }
        mustChangePassword: { type: boolean }
        lockedOut: { type: boolean, description: Derived — now < LockoutUntil }
        lastSignInAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
      required: [id, username, displayName, role, isActive, createdAt]
    UserCreate:
      type: object
      properties:
        username: { type: string, minLength: 3, maxLength: 64, pattern: '^[A-Za-z0-9._-]+$' }
        displayName: { type: string, minLength: 1, maxLength: 100 }
        role: { type: string, enum: [Admin, Employee] }
        password: { type: string, format: password, minLength: 10 }
      required: [username, displayName, role, password]
    UserPatch:
      type: object
      description: Partial update — only these three fields accepted (400 on unknown members)
      properties:
        role: { type: string, enum: [Admin, Employee] }
        isActive: { type: boolean, description: false ⇒ refuses sign-in AND kills live sessions next request }
        displayName: { type: string, minLength: 1, maxLength: 100 }
...
    Error:
      type: object
      properties:
        error: { type: string, example: unauthorized }
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized — missing/invalid/expired credential, or inactive account (generic)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden — Employee cannot perform Admin-only actions (upload/delete/people management)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'